delete obsolete node-fetch polyfill
node.js has native fetch for a while now.

it just doesn't have native XMLHttpRequest.
