--- src/fetch.browser.js
@@ -1,12 +1,0 @@
-'use strict'
-
-/**
- * @typedef {globalThis.Headers} Headers
- * @typedef {globalThis.Request} Request
- * @typedef {globalThis.Response} Response
- */
-
-// use window.fetch if it is available, fall back to node-fetch if not
-const nativeFetch = require('native-fetch')
-
-module.exports = nativeFetch
